Allow use of ${terraform.workspace} in tests
I'm working on a plugin which checks that resource names include
`${terraform.workspace}` (if it's set) -
https://github.com/richardTowers/tflint-ruleset-workspaces/

Currently it's not possible to test this functionality, because (unlike
tflint) tflint-plugin-sdk doesn't support the `${terraform.workspace}`
expression.

Fixing https://github.com/terraform-linters/tflint-plugin-sdk/issues/64
would be nicer, but as this is quite self contained I think it's okay to
add this feature to the TestRunner.